refactor: 拆分时间和背景视频同步状态桥hooks
